Improve JSON parsing performance
For scout results, there a lot of rows and columns of individual JSON to parse. This amplifies any overhead associated with use of the web worker for JSON parsing. A few fixes here:

1) Improve the web worker to send the JSON5 bundle only a single time
2) Other minor tweaks to web worker perf
3) The asyncJsonParse function is now smarter about parsing right on the current thread rather than delegating to the web worker. For small parses that will be very fast it doesn’t bother using the web worker
4) Our parsing logic for converting columntables to objects was sometimes sending null through to the parser, which it should (it was just doing `as string` which would let nulls through the type system)

In my testing, this yields approximately 4-5x faster loads for scanner dataframes (especially as they have high hundreds to thousands of rows).